test.sh: display an error if the one required argument is missing.
I just attempted to run ./test.sh without knowing what it did, and
was greeted with the following output:

  $ ./test.sh
  Using 15 as global script timeout
  Checking -t (-t) on 37 server(s)
  not listed in -t
  not listed in -t
  not listed in -t
  not listed in -t
  not listed in -t
  ...

Obviously that's not so good, and it would be better if the script
didn't try to make DNS requests when the user does something dumb.
This commit adds a check for the number of arguments, and will bail
out early if the required IP address argument is missing. Now, the
same incorrect invocation will produce,

  $ ./test.sh
  Usage: ./test.sh <ip address>

    <ip address> the IP to check against a bunch of blacklists.

Most importantly, it doesn't make any DNS requests in that case.